+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Minor: Allow for customizing the behavior of `Right Click`ing of usernames. (#4622, #4751)
 - Minor: Added support for opening incognito links in firefox-esr and chromium. (#4745)
 - Minor: Added support for opening incognito links under Linux/BSD using XDG. (#4745)
+- Minor: Add missing hotkey Action for Open Player in Browser. (#4756)
 - Bugfix: Increased amount of blocked users loaded from 100 to 1,000. (#4721)
 - Bugfix: Fixed generation of crashdumps by the browser-extension process when the browser was closed. (#4667)
 - Bugfix: Fix spacing issue with mentions inside RTL text. (#4677)